# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/customers/address": {
            "get": {
                "description": "Retorna un arreglo con todas las direcciones asociadas al cliente identificado en el token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers Address"
                ],
                "summary": "Obtiene todas las direcciones del cliente autenticado",
                "responses": {
                    "200": {
                        "description": "Lista de direcciones",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Address"
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "unexpected error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Registra una dirección asociada al customer_id obtenido del contexto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers Address"
                ],
                "summary": "Registra una nueva dirección para el cliente autenticado",
                "parameters": [
                    {
                        "description": "Datos de la dirección",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Address"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "address registred",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error de validación o parseo",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "unexpected error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/address/{address}": {
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Elimina la dirección especificada por su ID, siempre que pertenezca al cliente autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers Address"
                ],
                "summary": "Elimina una dirección del cliente autenticado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la dirección (UUID)",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "address deleted success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "unexpected error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/shops": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "End Point para obtener un negocio con todos sus products, se debe enviar alguno de los 2 query params requeridos.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Este es el ID del negocio, viene en formato UUID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Este es el tag del negocio .ej:dmo",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Latitud del usuario .ej: 4.678034",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitud del usuario .ej: -74.0496399",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ShopResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/shops/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Para que no me fastidies de cual es la ruta",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitud del usuario .ej: 4.678034",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitud del usuario .ej: -74.0496399",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ShopResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Address": {
            "type": "object",
            "required": [
                "address",
                "alias",
                "coords",
                "reference"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Calle 123 # 45-67"
                },
                "alias": {
                    "type": "string",
                    "example": "Casa principal"
                },
                "coords": {
                    "$ref": "#/definitions/dto.Coords"
                },
                "id": {
                    "type": "string"
                },
                "reference": {
                    "type": "string",
                    "example": "Frente al parque de los niños"
                }
            }
        },
        "dto.Coords": {
            "type": "object",
            "required": [
                "lat",
                "long"
            ],
            "properties": {
                "lat": {
                    "type": "number",
                    "example": 4.60971
                },
                "long": {
                    "type": "number",
                    "example": -74.08175
                }
            }
        },
...
        "dtos.ShopResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "banner": {
                    "type": "string"
                },
                "home_phone": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "opened": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                },
                "tag": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}